refactor : userSetting에서 Param방식에서 RequestBody형식으로 변경
