+++ src/main/java/com/example/thedayoftoday/domain/dto/user/PasswordUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.example.thedayoftoday.domain.dto.user;
+
+public record PasswordUpdateRequest(String newPassword) {
+}
